Moved createGenericMsg to AssetBase
