+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.wire</groupId>
     <artifactId>xenon</artifactId>
-    <version>1.5.4</version>
+    <version>1.5.5</version>
 
     <name>Xenon</name>
     <description>Base Wire Bots Library</description>
