+++ src/main/java/com/wire/xenon/assets/AssetBase.java
@@ -1,6 +1,8 @@
 package com.wire.xenon.assets;
 
 
+import com.google.protobuf.ByteString;
+import com.waz.model.Messages;
 import com.wire.xenon.tools.Util;
 
 import java.security.MessageDigest;
@@ -20,6 +22,7 @@ public abstract class AssetBase implements IAsset, IGeneric {
     protected String retention = "persistent";
     protected boolean readReceiptsEnabled = true;
     protected String domain;
+    protected long expires;
 
     public AssetBase(UUID messageId, String mimeType) {
         this.messageId = messageId;
@@ -37,6 +40,48 @@ public abstract class AssetBase implements IAsset, IGeneric {
 
         encBytes = Util.encrypt(otrKey, bytes, iv);
         sha256 = MessageDigest.getInstance("SHA-256").digest(encBytes);
+    }
+
+    @Override
+    public Messages.GenericMessage createGenericMsg() {
+        // Remote
+        Messages.Asset.RemoteData.Builder remote = Messages.Asset.RemoteData.newBuilder()
+                .setOtrKey(ByteString.copyFrom(getOtrKey()))
+                .setSha256(ByteString.copyFrom(getSha256()));
+
+        // Only set token on private assets
+        if (getAssetToken() != null) {
+            remote.setAssetToken(getAssetToken());
+        }
+
+        if (getAssetKey() != null) {
+            remote.setAssetId(getAssetKey());
+        }
+
+        if (getDomain() != null) {
+            remote.setAssetDomain(getDomain());
+        }
+
+        Messages.Asset.Builder asset = Messages.Asset.newBuilder()
+                .setExpectsReadConfirmation(isReadReceiptsEnabled())
+                .setUploaded(remote);
+
+        Messages.GenericMessage.Builder ret = Messages.GenericMessage.newBuilder()
+                .setMessageId(getMessageId().toString());
+
+        if (expires > 0) {
+            Messages.Ephemeral.Builder ephemeral = Messages.Ephemeral.newBuilder()
+                    .setAsset(asset)
+                    .setExpireAfterMillis(expires);
+
+            return ret
+                    .setEphemeral(ephemeral)
+                    .build();
+        }
+
+        return ret
+                .setAsset(asset)
+                .build();
     }
 
     @Override
@@ -116,11 +161,21 @@ public abstract class AssetBase implements IAsset, IGeneric {
         this.sha256 = sha256;
     }
 
+    @Override
     public String getDomain() {
         return domain;
     }
 
     public void setDomain(String domain) {
         this.domain = domain;
+    }
+
+    @Override
+    public long getExpires() {
+        return expires;
+    }
+
+    public void setExpires(long expires) {
+        this.expires = expires;
     }
 }
