+++ src/main/java/com/wire/xenon/assets/AudioAsset.java
@@ -32,30 +32,4 @@ public class AudioAsset extends AssetBase {
     public AudioAsset(byte[] bytes, AudioPreview preview) throws Exception {
         super(preview.getMessageId(), preview.getMimeType(), bytes);
     }
-
-    @Override
-    public Messages.GenericMessage createGenericMsg() {
-        Messages.Asset.RemoteData.Builder remote = Messages.Asset.RemoteData.newBuilder()
-                .setOtrKey(ByteString.copyFrom(getOtrKey()))
-                .setSha256(ByteString.copyFrom(getSha256()));
-
-        // Only set token on private assets
-        if (getAssetToken() != null) {
-            remote.setAssetToken(getAssetToken());
-        }
-
-        if (getAssetKey() != null) {
-            remote.setAssetId(getAssetKey());
-        }
-
-        Messages.Asset asset = Messages.Asset.newBuilder()
-                .setUploaded(remote.build())
-                .setExpectsReadConfirmation(isReadReceiptsEnabled())
-                .build();
-
-        return Messages.GenericMessage.newBuilder()
-                .setMessageId(getMessageId().toString())
-                .setAsset(asset)
-                .build();
-    }
 }
